% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Best_Assingment.R
\name{best_hit_taxonomy_assignment}
\alias{best_hit_taxonomy_assignment}
\title{Hierarchical best-hit taxonomy assignment with per-rank fallback rule}
\usage{
best_hit_taxonomy_assignment(blast_qc, cutoffs_long, defaults)
}
\description{
Pass ONLY those OTUs that haven't been assigned already!
For each rank, if the best e-value hit is undefined and the second-best hit is defined
and at least 60% as good, use the second-best hit's value for that rank.
}
