% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_N.R
\name{check_N}
\alias{check_N}
\title{Check proportion of N bases in each sequence.}
\usage{
check_N(rep_seqs, cutoff = 1)
}
\arguments{
\item{rep_seqs}{DNAStringSet or character vector of representative sequences.}

\item{cutoff}{Numeric, percent threshold (default 1).}
}
\value{
Data frame with columns: qseqid, N_percent, N_flag.
}
\description{
Calculates the proportion of "N" bases (ambiguous bases) in each sequence and flags if above the given threshold.
}
\examples{
seqs <- Biostrings::DNAStringSet(c(seq1 = "ATGCNNNN", seq2 = "NNNNATGC"))
check_N(seqs)
check_N(seqs, cutoff = 10)
}
