% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concensus_taxonomy.R
\name{consensus_taxonomy_assignment}
\alias{consensus_taxonomy_assignment}
\title{Per-rank consensus filter for taxonomy assignment}
\usage{
consensus_taxonomy_assignment(final_table, blast_qc)
}
\arguments{
\item{final_table}{Data frame of taxonomic assignments.}

\item{blast_qc}{Data frame of filtered BLAST hits for each OTU.}
}
\value{
Data frame of consensus assignments (same structure as input).
}
\description{
Only confirms or demotes, never promotes Unclassified.
}
