% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{load_and_check}
\alias{load_and_check}
\title{Load and check BLAST results and rep-seq FASTA}
\usage{
load_and_check(blast_file, rep_fasta, taxonomy_col = "stitle")
}
\arguments{
\item{blast_file}{Path to BLAST results TSV file}

\item{rep_fasta}{Path to representative sequences FASTA file}

\item{taxonomy_col}{The column in BLAST file containing taxonomy strings (default "stitle")}
}
\value{
List with BLAST dataframe (kingdom-filtered) and rep_seqs DNAStringSet
}
\description{
Load and check BLAST results and rep-seq FASTA
}
