% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concensus_taxonomy.R
\name{safe_rbind_list}
\alias{safe_rbind_list}
\title{Safely rbinds list of data frames, ensuring columns match}
\usage{
safe_rbind_list(dfs, all_cols = NULL)
}
\arguments{
\item{dfs}{List of data frames}

\item{all_cols}{Vector of required columns}
}
\value{
Combined data frame
}
\description{
Safely rbinds list of data frames, ensuring columns match
}
