% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concensus_taxonomy.R
\name{write_initial_assignments}
\alias{write_initial_assignments}
\title{Create and write the initial assignments table including drops at all steps}
\usage{
write_initial_assignments(
  easy_df,
  consensus_df,
  rep_seqs,
  blast,
  blast_filtered,
  file = "outputs/initial_assignments.csv"
)
}
\arguments{
\item{easy_df}{Data frame of easy-assigned OTUs}

\item{consensus_df}{Data frame of consensus-assigned OTUs (hard ones)}

\item{rep_seqs}{DNAStringSet or named character vector of rep seqs}

\item{blast}{Data frame of all BLAST results}

\item{blast_filtered}{Data frame of filtered BLAST results}

\item{file}{Path for output CSV}
}
\value{
Data frame written to CSV (and for downstream analysis)
}
\description{
Create and write the initial assignments table including drops at all steps
}
