% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concensus_taxonomy.R
\name{ensure_cols}
\alias{ensure_cols}
\title{Ensure data frame has all required columns (as character)}
\usage{
ensure_cols(df, all_cols)
}
\arguments{
\item{df}{Data frame to fix}

\item{all_cols}{Vector of required columns}
}
\value{
Fixed data frame (in correct order, with all columns present)
}
\description{
Ensure data frame has all required columns (as character)
}
