% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{trim_alignments}
\alias{trim_alignments}
\title{Trim BLAST alignments by minimum length (default fraction: 0.6 of median rep-seq length)}
\usage{
trim_alignments(blast, rep_seqs, fraction = 0.6)
}
\arguments{
\item{blast}{BLAST dataframe}

\item{rep_seqs}{DNAStringSet}

\item{fraction}{Numeric fraction of median rep-seq length (default: 0.6)}
}
\value{
Filtered BLAST dataframe
}
\description{
Trim BLAST alignments by minimum length (default fraction: 0.6 of median rep-seq length)
}
